requirements.txt
colour_demosaicing==0.2.5
einops==0.4.1
ExifRead==3.0.0
lpips==0.1.4
matplotlib==3.5.3
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
Pillow==10.3.0
pyiqa==0.1.11
rawpy==0.18.1
scikit_image==0.19.3
scipy==1.9.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
timm==0.9.7
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
